rag-mbe-agent/requirements.txt
# ── Core Framework ────────────────────────────────────────────────────────────
langgraph==0.2.62
langchain==0.3.14
langchain-community==0.3.14
langchain-ollama==0.2.3
langchain-text-splitters==0.3.4
# ❗ dejar que se resuelva solo (clave)
langchain-core>=0.3.33,<0.4.0

# ── API ───────────────────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10

# ── Embeddings & ML ───────────────────────────────────────────────────────────
transformers==4.47.1
# ⚠️ importante en Mac (evita problemas de build)
torch==2.5.1
sentence-transformers==3.3.1
huggingface-hub==0.27.0

# ── Vector Store ──────────────────────────────────────────────────────────────
faiss-cpu==1.9.0
numpy==1.26.4

# ── Document Loading ──────────────────────────────────────────────────────────
pypdf==5.1.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
tenacity==9.0.0
httpx==0.28.1
typing-extensions==4.12.2
